default tensor type

+++ RBM/rbm.py
@@ -17,7 +17,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-torch.set_default_tensor_type(float) 
+torch.set_default_tensor_type(torch.float64) 
 # Float in python is equal to double in C
 # https://docs.python.org/3/library/stdtypes.html
